requirements.txt
Flask
bcrypt
google-genai
Flask-Dance